Fix fields not saving their values when disabled
We grey out some fields because we don't want to allow the user to
change them. However, we were setting the "disabled" jQuery property to
do this, and this results in the read value for the form field being
blank and thus the field in the database being blank. Set the "readonly"
property instead to get the correct behaviour.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>