Remove milestone name from recipe details page
The idea of having the milestone name showing in the
h1 of the recipe details page was based on the assumption
that there would a recipe details page per recipe
per milestone. This is not the case: there is only one
recipe details page that shows all the existing updates
up to the current milestone.

In this situation, if you land on the recipe details page
from a past milestone, the display of the current milestone
is confusing, since it doesn't match the milestone you
came from.

An easy way to sort this out is simply remove the milestone
from the page heading.

Signed-off-by: Belen Barros Pena <belen.barros.pena@intel.com>