README: update setup instructions
With Django 1.8 you don't need to use syncdb since migrate does all of
the database structure setup (and in fact in 1.9 it has been removed -
we don't support that version yet but it's good to avoid that issue in
future). The only other thing that syncdb did was to create a superuser,
and there is now a specific command to do that. Since you do need to
actually log in as part of some of the later steps, we now tell the user
explicitly to run that in the first step.

Additionally, add a section about what to do when upgrading - normally
this is straightforward, but if you're upgrading from prior to the
Django 1.8 move then you need to fake the initial migration or it'll
fail (since the structure is already present).

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>