detail: Fix layer description not wrapping properly
In 3adddf6a25977cb576f8fcefd307d947e43d637b I changed the style to allow
pre-formatted text to work in the layer description, but this broke
wrapping so that text could go behind the "Dependencies" box. Use the
"pre-wrap" style instead so that both pre-formatting and wrapping work
here.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>